Remove isPassed as abstract function from Scheincriteria.
This is due to the facts that:

a) Most code in isPassed & checkCriteriaStatus would have been duplicate code OR wired in strange ways.

b) isPassed is never used outside the classes (or is never used in general depending in the class).

c) If there are helper methods needed on can implement them on a per class basis and with private restrictions.