Fix GradingDocuments not being saved as refs.
Due to the mapProps decorator not allowing document references the StudentDocument now holds an actual map (which does not get saved to the DB) and an array only with the documents (which gets used for the refs).

With this approach, the GradingDocuments can be used in multiple students at once.