Fix several things regarding invalid intervals.
SelectInterval: Fix not propagating invalid intervals properly if start or end date are "null". If one of those are "null" an invalid Interval with a corresponding message is created and set as value.

WeekdayBox: Sorting failed if one of the intervals where invalid. Now, invalid intervals get sorted to the top of the list.