Update to TS 4.0 (beta) - more info below.
The update is done bc the CustomRoute.combin() function needs the new spread functionalities to properly work. Due to the fact that the release of v2 is scheduled after the release of TS 4.0 the beta can be used right now.

This also fixes some type issues with the new version.