Add proper database connection handling on server start.
Tries to connect to the database when the server starts. If it can't connect, it'll  throw an error. If it can the SQLDatabaseModule ensures that

a) the database exists (if not it gets created).

b) the schema of the database matches the required one (if not it gets safely updated).